+++ Module1/pom.xml
@@ -4,8 +4,8 @@
 		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>module.1</groupId>
-	<artifactId>module.1</artifactId>
+	<groupId>com.epam.olukash</groupId>
+	<artifactId>module1</artifactId>
 	<version>1.0-SNAPSHOT</version>
