+++ Module1/src/main/java/com/epam/olukash/module1/correct/bean/Currency.java
@@ -1,0 +1,9 @@
+package com.epam.olukash.module1.correct.bean;
+
+/**
+ * @author Oleksii_Lukash
+ * @date 11/28/2015
+ */
+public interface Currency
+{
+}
