+++ Module1/src/main/java/com/epam/olukash/module1/correct/bean/Mortgage.java
@@ -1,0 +1,14 @@
+package com.epam.olukash.module1.correct.bean;
+
+/**
+ * @author Oleksii_Lukash
+ * @date 11/28/2015
+ */
+public class Mortgage extends AbstractLoanAccount
+{
+	@Override
+	protected long getAccountRate()
+	{
+		return 0;
+	}
+}
