+++ Module1/src/main/java/com/epam/olukash/module1/correct/manager/CreditCardManagerImpl.java
@@ -1,0 +1,43 @@
+package com.epam.olukash.module1.correct.manager;
+
+import java.math.BigDecimal;
+import java.util.List;
+
+import com.epam.olukash.module1.correct.bean.CreditCard;
+
+/**
+ * @author Oleksii_Lukash
+ * @date 11/28/2015
+ */
+public class CreditCardManagerImpl implements CreditCardManager
+{
+	public List<CreditCard> getVisaCreditCards()
+	{
+		return null;
+	}
+
+	public BigDecimal getAmountOfLastTransaction(Long creditCardID)
+	{
+		return null;
+	}
+
+	public void save(CreditCard bean)
+	{
+
+	}
+
+	public void update(CreditCard bean)
+	{
+
+	}
+
+	public CreditCard get(long id)
+	{
+		return null;
+	}
+
+	public void remove(CreditCard bean)
+	{
+
+	}
+}
