+++ Module1/src/main/java/com/epam/olukash/module1/correct/manager/MortgageManagerImpl.java
@@ -1,0 +1,37 @@
+package com.epam.olukash.module1.correct.manager;
+
+import java.util.List;
+
+import com.epam.olukash.module1.correct.bean.Mortgage;
+
+/**
+ * @author Oleksii_Lukash
+ * @date 11/28/2015
+ */
+public class MortgageManagerImpl implements MortgageManager
+{
+	public List<Mortgage> getBadGuys()
+	{
+		return null;
+	}
+
+	public void save(Mortgage bean)
+	{
+
+	}
+
+	public void update(Mortgage bean)
+	{
+
+	}
+
+	public Mortgage get(long id)
+	{
+		return null;
+	}
+
+	public void remove(Mortgage bean)
+	{
+
+	}
+}
