+++ Module1/src/main/java/com/epam/olukash/module1/correct/manager/UserManager.java
@@ -1,0 +1,11 @@
+package com.epam.olukash.module1.correct.manager;
+
+import com.epam.olukash.module1.correct.bean.User;
+
+/**
+ * @author Oleksii_Lukash
+ * @date 11/28/2015
+ */
+public interface UserManager extends AbstractManager<User>
+{
+}
