+++ Module1/src/main/java/com/epam/olukash/module1/correct/manager/UserManagerImpl.java
@@ -1,0 +1,30 @@
+package com.epam.olukash.module1.correct.manager;
+
+import com.epam.olukash.module1.correct.bean.User;
+
+/**
+ * @author Oleksii_Lukash
+ * @date 11/28/2015
+ */
+public class UserManagerImpl implements UserManager
+{
+	public void save(User bean)
+	{
+
+	}
+
+	public void update(User bean)
+	{
+
+	}
+
+	public User get(long id)
+	{
+		return null;
+	}
+
+	public void remove(User bean)
+	{
+
+	}
+}
