+++ Module1/src/main/java/com/epam/olukash/module1/erroneous/bean/ErroneousCurrencyUSD.java
@@ -1,0 +1,9 @@
+package com.epam.olukash.module1.erroneous.bean;
+
+/**
+ * @author Oleksii_Lukash
+ * @date 11/28/2015
+ */
+public class ErroneousCurrencyUSD
+{
+}
