+++ Module1/src/main/java/com/epam/olukash/module1/erroneous/manager/ErroneousAccountManagerImpl.java
@@ -1,0 +1,44 @@
+package com.epam.olukash.module1.erroneous.manager;
+
+import java.util.List;
+
+import com.epam.olukash.module1.erroneous.bean.ErroneousCreditCard;
+import com.epam.olukash.module1.erroneous.bean.ErroneousMortgage;
+import com.epam.olukash.module1.erroneous.bean.ErroneousUser;
+
+/**
+ * @author Oleksii_Lukash
+ * @date 11/28/2015
+ */
+public class ErroneousAccountManagerImpl implements ErroneousAccountManager
+{
+	public ErroneousCreditCard saveDepositAccount(ErroneousCreditCard depositAccount)
+	{
+		return null;
+	}
+
+	public ErroneousCreditCard updateDepositAccount(ErroneousCreditCard depositAccount)
+	{
+		return null;
+	}
+
+	public ErroneousMortgage saveOverdraftAccount(ErroneousMortgage overdraftAccount)
+	{
+		return null;
+	}
+
+	public ErroneousMortgage updateOverdraftAccount(ErroneousMortgage overdraftAccount)
+	{
+		return null;
+	}
+
+	public void updateUser(ErroneousUser user)
+	{
+
+	}
+
+	public List<ErroneousUser> getAllUsers()
+	{
+		return null;
+	}
+}
