--- Module1/src/main/java/oleksii/lukash/module1/correct/bean/AbstractBean.java
@@ -1,34 +1,0 @@
-package oleksii.lukash.module1.correct.bean;
-
-import java.util.Date;
-
-/**
- * Created by Oleksii_Lukash on 11/28/2015.
- */
-public abstract class AbstractBean
-{
-	private long ID;
-	private Date createdDate;
-
-	//////////////////////////////////////////////////////
-
-	public Date getCreatedDate()
-	{
-		return createdDate;
-	}
-
-	public void setCreatedDate(Date createdDate)
-	{
-		this.createdDate = createdDate;
-	}
-
-	public long getID()
-	{
-		return ID;
-	}
-
-	public void setID(long ID)
-	{
-		this.ID = ID;
-	}
-}
