--- Module1/src/main/java/oleksii/lukash/module1/correct/bean/Currency.java
@@ -1,8 +1,0 @@
-package oleksii.lukash.module1.correct.bean;
-
-/**
- * Created by Oleksii_Lukash on 11/28/2015.
- */
-public interface Currency
-{
-}
