--- Module1/src/main/java/oleksii/lukash/module1/correct/bean/CurrencyUSD.java
@@ -1,8 +1,0 @@
-package oleksii.lukash.module1.correct.bean;
-
-/**
- * Created by Oleksii_Lukash on 11/28/2015.
- */
-public class CurrencyUSD extends Currency
-{
-}
