--- Module1/src/main/java/oleksii/lukash/module1/correct/bean/Mortgage.java
@@ -1,13 +1,0 @@
-package oleksii.lukash.module1.correct.bean;
-
-/**
- * Created by Oleksii_Lukash on 11/28/2015.
- */
-public class Mortgage extends AbstractLoanAccount
-{
-	@Override
-	protected long getAccountRate()
-	{
-		return 0;
-	}
-}
