--- Module1/src/main/java/oleksii/lukash/module1/correct/manager/CreditCardManagerImpl.java
@@ -1,42 +1,0 @@
-package oleksii.lukash.module1.correct.manager;
-
-import java.math.BigDecimal;
-import java.util.List;
-
-import oleksii.lukash.module1.correct.bean.CreditCard;
-
-/**
- * Created by Oleksii_Lukash on 11/28/2015.
- */
-public class CreditCardManagerImpl implements CreditCardManager
-{
-	public List<CreditCard> getVisaCreditCards()
-	{
-		return null;
-	}
-
-	public BigDecimal getAmountOfLastTransaction()
-	{
-		return null;
-	}
-
-	public CreditCard save(CreditCard bean)
-	{
-		return null;
-	}
-
-	public CreditCard update(CreditCard bean)
-	{
-		return null;
-	}
-
-	public CreditCard get(long id)
-	{
-		return null;
-	}
-
-	public void remove(CreditCard bean)
-	{
-
-	}
-}
