--- Module1/src/main/java/oleksii/lukash/module1/correct/manager/MortgageManagerImpl.java
@@ -1,36 +1,0 @@
-package oleksii.lukash.module1.correct.manager;
-
-import java.util.List;
-
-import oleksii.lukash.module1.correct.bean.Mortgage;
-
-/**
- * Created by Oleksii_Lukash on 11/28/2015.
- */
-public class MortgageManagerImpl implements MortgageManager
-{
-	public List<Mortgage> getBadGuys()
-	{
-		return null;
-	}
-
-	public Mortgage save(Mortgage bean)
-	{
-		return null;
-	}
-
-	public Mortgage update(Mortgage bean)
-	{
-		return null;
-	}
-
-	public Mortgage get(long id)
-	{
-		return null;
-	}
-
-	public void remove(Mortgage bean)
-	{
-
-	}
-}
