--- Module1/src/main/java/oleksii/lukash/module1/correct/manager/UserManager.java
@@ -1,8 +1,0 @@
-package oleksii.lukash.module1.correct.manager;
-
-/**
- * Created by Oleksii_Lukash on 11/28/2015.
- */
-public interface UserManager extends AbstractManager
-{
-}
