--- Module1/src/main/java/oleksii/lukash/module1/correct/manager/UserManagerImpl.java
@@ -1,27 +1,0 @@
-package oleksii.lukash.module1.correct.manager;
-
-/**
- * Created by Oleksii_Lukash on 11/28/2015.
- */
-public class UserManagerImpl implements UserManager
-{
-	public Object save(Object bean)
-	{
-		return null;
-	}
-
-	public Object update(Object bean)
-	{
-		return null;
-	}
-
-	public Object get(long id)
-	{
-		return null;
-	}
-
-	public void remove(Object bean)
-	{
-
-	}
-}
