--- Module1/src/main/java/oleksii/lukash/module1/erroneous/bean/ErroneousCurrencyUSD.java
@@ -1,8 +1,0 @@
-package oleksii.lukash.module1.erroneous.bean;
-
-/**
- * Created by Oleksii_Lukash on 11/28/2015.
- */
-public class ErroneousCurrencyUSD
-{
-}
