--- Module1/src/main/java/oleksii/lukash/module1/erroneous/manager/ErroneousAccountManagerImpl.java
@@ -1,43 +1,0 @@
-package oleksii.lukash.module1.erroneous.manager;
-
-import java.util.List;
-
-import oleksii.lukash.module1.erroneous.bean.ErroneousCreditCard;
-import oleksii.lukash.module1.erroneous.bean.ErroneousMortgage;
-import oleksii.lukash.module1.erroneous.bean.ErroneousUser;
-
-/**
- * Created by Oleksii_Lukash on 11/28/2015.
- */
-public class ErroneousAccountManagerImpl implements ErroneousAccountManager
-{
-	public ErroneousCreditCard saveDepositAccount(ErroneousCreditCard depositAccount)
-	{
-		return null;
-	}
-
-	public ErroneousCreditCard updateDepositAccount(ErroneousCreditCard depositAccount)
-	{
-		return null;
-	}
-
-	public ErroneousMortgage saveOverdraftAccount(ErroneousMortgage overdraftAccount)
-	{
-		return null;
-	}
-
-	public ErroneousMortgage updateOverdraftAccount(ErroneousMortgage overdraftAccount)
-	{
-		return null;
-	}
-
-	public void updateUser(ErroneousUser user)
-	{
-
-	}
-
-	public List<ErroneousUser> getAllusers()
-	{
-		return null;
-	}
-}
